feat(MS-11): Add charts

+++ src/components/transactions/TransactionList/TransactionList.tsx
@@ -3,6 +3,8 @@ import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "../../../Store/store";
 import SinglePayment from "../SinglePayment";
 import { deleteSingleTransaction } from "../../slices/TransactionFormSlice";
+import { changeBalanceWithTransaction } from "../../slices/BudgetSlice";
+import { useMemo } from "react";
 
 type TransactionListTypes = {
 	onShowModal: () => void;
@@ -12,10 +14,28 @@ function TransactionList({ onShowModal }: TransactionListTypes) {
 	const transactions = useSelector(
 		(state: RootState) => state.transaction.transactions
 	);
+
+	const selectedCard = useSelector(
+		(state: RootState) => state.card.selectedCard
+	);
 	const dispatch = useDispatch();
 
+	const filteredTransactions = useMemo(() => {
+		if (!selectedCard) return [];
+		return transactions.filter((t) => t.cardId === selectedCard.id);
+	}, [transactions, selectedCard]);
+
 	const handleRemoveSingleTransaction = (id: string) => {
+		const transactionToDelete = transactions.find((t) => t.id === id);
+		if (!transactionToDelete || !selectedCard) return;
+
 		dispatch(deleteSingleTransaction({ id }));
+		dispatch(
+			changeBalanceWithTransaction({
+				cardId: selectedCard?.id,
+				transaction: transactionToDelete,
+			})
+		);
 	};
 	return (
 		<S.TransactionContainer>
@@ -27,7 +47,7 @@ function TransactionList({ onShowModal }: TransactionListTypes) {
 				<S.AddTransactionButton onClick={onShowModal} text="Add Transaction" />
 			</S.SubMenu>
 			<S.Ul>
-				{transactions.map((transaction) => (
+				{filteredTransactions.map((transaction) => (
 					<SinglePayment
 						key={transaction.id}
 						id={transaction.id}
